RF: functional changes implemented.
- Removed Renderer and Screen classes.
- Introduced methods to play in the show_manager.
