+++ fury/v2/window/__init__.pyi
@@ -1,5 +1,22 @@
-__all__ = ["ShowManager", "Scene", "record", "Renderer", "Screen"]
+__all__ = [
+    "ShowManager",
+    "Scene",
+    "record",
+    "reset_camera",
+    "update_camera",
+    "update_screens",
+    "render_screens",
+    "calculate_screen_sizes",
+    "create_screen",
+]
 
-from .renderer import Renderer, Screen
-from .scene import Scene
+from .containers import (
+    Scene,
+    calculate_screen_sizes,
+    create_screen,
+    render_screens,
+    reset_camera,
+    update_camera,
+    update_screens,
+)
 from .show_manager import ShowManager, record
