+++ fury/v2/window/show_manager.py
@@ -2,17 +2,20 @@ from functools import reduce
 
 from PIL.Image import fromarray as image_from_array
 from numpy import asarray as np_asarray
-from pygfx import DirectionalLight, OrbitController, PerspectiveCamera
+from pygfx import WgpuRenderer
 from wgpu.gui.auto import WgpuCanvas, run
 from wgpu.gui.jupyter import WgpuCanvas as JupyterWgpuCanvas
 from wgpu.gui.offscreen import WgpuCanvas as OffscreenWgpuCanvas
 
-from fury.decorators import warn_on_args_to_kwargs
-from fury.v2.window import Renderer, Screen
+from fury.v2.window import (
+    calculate_screen_sizes,
+    create_screen,
+    render_screens,
+    update_screens,
+)
 
 
 class ShowManager:
-    @warn_on_args_to_kwargs()
     def __init__(
         self,
         *,
@@ -35,39 +38,27 @@ class ShowManager:
         screen_config=None,
     ):
         self.size = size
-        if window_type == "auto":
-            self.window = WgpuCanvas(size=self.size, title=title)
-        elif window_type == "jupyter":
-            self.window = JupyterWgpuCanvas(size=self.size, title=title)
-        elif window_type == "offscreen":
-            self.window = OffscreenWgpuCanvas(size=self.size, title=title)
-        else:
-            self.window = WgpuCanvas(size=self.size, title=title)
+        self._title = title
+        self._setup_window(window_type)
 
         if renderer is None:
-            renderer = Renderer(self.window)
+            renderer = WgpuRenderer(self.window)
         self.renderer = renderer
 
-        self.scene = scene
         self._screen_config = screen_config
-
         self._total_screens = 0
+        self._calculate_total_screens()
 
-        if self._screen_config is None:
-            self._total_screens = 1
-        elif isinstance(self._screen_config[0], int):
-            self._total_screens = reduce(lambda a, b: a + b, self._screen_config)
-        else:
-            self._total_screens = len(self._screen_config)
-
+        self.scene = scene
         if not isinstance(scene, list):
             self.scene = [scene] * self._total_screens
 
-        self.renderer.screens = self._create_screens()
-        self.renderer.update_screens(
-            _calculate_screen_sizes(self._screen_config, self.renderer.logical_size)
+        self.screens = self._create_screens()
+        update_screens(
+            self.screens,
+            calculate_screen_sizes(self._screen_config, self.renderer.logical_size),
         )
-        self._title = title
+
         self.camera = camera
         self.png_magnify = png_magnify
         self.reset_camera = reset_camera
@@ -80,16 +71,32 @@ class ShowManager:
         self.renderer.pixel_ratio = pixel_ratio
         self.renderer.blend_mode = "weighted_plus"
 
-        # if controller_style == "orbit":
-        #     OrbitController(self.camera, register_events=self.renderer)
         self.renderer.enable_events()
 
         self.renderer.add_event_handler(self.resize, "resize")
 
+    def _setup_window(self, window_type):
+        if window_type == "auto":
+            self.window = WgpuCanvas(size=self.size, title=self._title)
+        elif window_type == "jupyter":
+            self.window = JupyterWgpuCanvas(size=self.size, title=self._title)
+        elif window_type == "offscreen":
+            self.window = OffscreenWgpuCanvas(size=self.size, title=self._title)
+        else:
+            self.window = WgpuCanvas(size=self.size, title=self._title)
+
+    def _calculate_total_screens(self):
+        if self._screen_config is None:
+            self._total_screens = 1
+        elif isinstance(self._screen_config[0], int):
+            self._total_screens = reduce(lambda a, b: a + b, self._screen_config)
+        else:
+            self._total_screens = len(self._screen_config)
+
     def _create_screens(self):
         screens = []
         for i in range(self._total_screens):
-            screens.append(Screen(self.renderer, scene=self.scene[i]))
+            screens.append(create_screen(self.renderer, scene=self.scene[i]))
         return screens
 
     @property
@@ -114,20 +121,20 @@ class ShowManager:
         img.save(fname)
 
     def render(self):
-        self.window.request_draw(lambda: self.renderer.render_screens())
+        self.window.request_draw(lambda: render_screens(self.renderer, self.screens))
 
     def start(self):
         self.render()
         run()
 
-    def resize(self, event):
-        self.renderer.update_screens(
-            _calculate_screen_sizes(self._screen_config, self.renderer.logical_size)
+    def resize(self, _event):
+        update_screens(
+            self.screens,
+            calculate_screen_sizes(self._screen_config, self.renderer.logical_size),
         )
         self.render()
 
 
-@warn_on_args_to_kwargs()
 def record(*, renderer=None, scene=None, fname="output.png"):
     if renderer is None:
         show_m = ShowManager(scene=scene, window_type="offscreen")
@@ -136,29 +143,3 @@ def record(*, renderer=None, scene=None, fname="output.png"):
         show_m.snapshot(fname)
     else:
         show_m = ShowManager(renderer=renderer, window_type="offscreen")
-
-
-def _calculate_screen_sizes(screens, size):
-    if screens is None:
-        return [(0, 0, *size)]
-
-    screen_bbs = []
-
-    v_sections = len(screens)
-    width = (1 / v_sections) * size[0]
-    x = 0
-
-    for h_section in screens:
-        if h_section == 0:
-            continue
-
-        height = (1 / h_section) * size[1]
-        y = 0
-
-        for _ in range(h_section):
-            screen_bbs.append((x, y, width, height))
-            y += height
-
-        x += width
-
-    return screen_bbs
