+++ requirements/default.txt
@@ -8,3 +8,5 @@ aiohttp>=3.8.4
 pygltflib>=1.15.1
 lazy_loader>=0.4
 pygfx>=0.3.0
+glfw>=2.7.0
+jupyter-rfb>=0.4.4
